Add files via upload
Scientific calculators are used widely in situations that require quick access to certain mathematical functions, especially those that were once looked up in mathematical tables, such as trigonometric functions or logarithms. They are also used for calculations of very large or very small numbers, as in some aspects of astronomy, physics, and chemistry.
They are very often required for math classes from the junior high school level through college, and are generally either permitted or required on many standardized tests covering math and science subjects; as a result, many are sold into educational markets to cover this demand, and some high-end models include features making it easier to translate a problem on a textbook page into calculator input, e.g. by providing a method to enter an entire problem in as it is written on the page using simple formatting tools.
	This is a calculator that resembles a typical scientific calculator, albeit a simpler version. In our version, we have only included the trigonometric functions and the logarithmic functions. The reason of creating a simpler version of the calculator is to help users to learn the programming concepts in a gradual manner and not to confuse them especially those who are learning to program in Visual Basic.
To design the interface, we just to need to modify the interface of the basic calculator that we have created earlier using Visual Basic 6.  In this calculator, we have added five more buttons, they are Sin, Cos, Tan, Log and Ln.  The common trigonometric functions in Visual Basic 6 are Sin, Cos, Tan.